firecracker: Set --enable-pci option by default when available
This enables the PCI virtio transport which is supposed to be faster than the MMIO one.

Support for this flag is currently only availble in Nixpkgs unstable.

I haven't been able to measure these perf improvements.

However this is still what the Firecracker docs say to do, so I guess let's make it the default?