firecracker: Add extraConfig option
For experimentation with low-level settings, add support for merging in
arbitrary options to the Firecracker config file, following the standard pattern
in nixpkgs for this kind of thing.

My immediate usecase for this is for experimenting with firecracker features
that are not yet released, so it doesn't make sense to add more specific Nix
options for those features.